스키마와 학생 엔티티 타입, 필드 불일치 수정

+++ src/main/java/com/example/quick_gg/entity/RefreshTokenEntity.java
@@ -15,5 +15,5 @@ public class RefreshTokenEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long
+    private Integer
 }
